requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.2
black==23.11.0
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.0
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
dm-tree==0.1.8
dnspython==2.4.2
executing==2.0.1
Flask==3.0.2
flatbuffers==24.3.7
fonttools==4.42.1
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
gspread==6.0.1
h11==0.14.0
h5py==3.10.0
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==3.0.5
kiwisolver==1.4.5
libclang==16.0.6
lstm==0.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mypy-extensions==1.0.0
namex==0.0.7
nest-asyncio==1.6.0
nodeenv==1.8.0
numpy==1.25.2
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.1
pandas==2.1.3
parso==0.8.3
pathspec==0.11.2
patsy==0.5.6
Pillow==10.0.0
platformdirs==4.0.0
plotly==5.18.0
prompt-toolkit==3.0.40
protobuf==4.25.3
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
Pygments==2.16.1
pymongo==4.6.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.7.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
StrEnum==0.4.15
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
tornado==6.3.3
traitlets==5.13.0
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.9
webdriver-manager==4.0.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
zipp==3.17.0
